feat(frontend): small refactor and move the pages inside folders with separated mocks and utils
TICKET-0000
