+++ frontend/src/pages/Alterts/Alerts.tsx
@@ -1,119 +1,14 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Navigation } from '@/components/Navigation'
-
-type Alert = {
-  id: string
-  type: 'risk' | 'opportunity' | 'news' | 'performance'
-  title: string
-  message: string
-  timestamp: Date
-  isRead: boolean
-  severity: 'low' | 'medium' | 'high'
-}
-
-const mockAlerts: Alert[] = [
-  {
-    id: '1',
-    type: 'risk',
-    title: 'High Volatility Detected',
-    message: 'Tesla (TSLA) showing increased volatility. Consider reviewing position size.',
-    timestamp: new Date(Date.now() - 1000 * 60 * 30), // 30 mins ago
-    isRead: false,
-    severity: 'high'
-  },
-  {
-    id: '2',
-    type: 'opportunity',
-    title: 'Rebalancing Opportunity',
-    message: 'Your tech allocation is 15% over target. Consider taking profits.',
-    timestamp: new Date(Date.now() - 1000 * 60 * 60 * 2), // 2 hours ago
-    isRead: false,
-    severity: 'medium'
-  },
-  {
-    id: '3',
-    type: 'news',
-    title: 'Market Update',
-    message: 'Fed meeting minutes released. Potential impact on your bond positions.',
-    timestamp: new Date(Date.now() - 1000 * 60 * 60 * 4), // 4 hours ago
-    isRead: true,
-    severity: 'low'
-  },
-  {
-    id: '4',
-    type: 'performance',
-    title: 'Portfolio Milestone',
-    message: 'Congratulations! Your portfolio hit a new all-time high.',
-    timestamp: new Date(Date.now() - 1000 * 60 * 60 * 24), // 1 day ago
-    isRead: true,
-    severity: 'low'
-  },
-  {
-    id: '5',
-    type: 'opportunity',
-    title: 'Dividend Alert',
-    message: 'Microsoft (MSFT) ex-dividend date approaching. Consider position timing.',
-    timestamp: new Date(Date.now() - 1000 * 60 * 60 * 24 * 2), // 2 days ago
-    isRead: true,
-    severity: 'medium'
-  }
-]
+import { mockAlerts } from './mocks'
+import { Alert } from './interfaces'
+import { formatTimestamp, getAlertIcon, getSeverityColor } from './utils'
 
 export default function Alerts() {
   const { t } = useTranslation()
   const [alerts, setAlerts] = useState<Alert[]>(mockAlerts)
   const [filter, setFilter] = useState<'all' | 'unread' | 'risk' | 'opportunity'>('all')
-
-  const getAlertIcon = (type: Alert['type']) => {
-    switch (type) {
-      case 'risk':
-        return (
-          <svg className="w-5 h-5 text-destructive" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.962-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z" />
-          </svg>
-        )
-      case 'opportunity':
-        return (
-          <svg className="w-5 h-5 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 7h8m0 0v8m0-8l-8 8-4-4-6 6" />
-          </svg>
-        )
-      case 'news':
-        return (
-          <svg className="w-5 h-5 text-blue-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 20H5a2 2 0 01-2-2V6a2 2 0 012-2h10a2 2 0 012 2v1m2 13a2 2 0 01-2-2V7m2 13a2 2 0 002-2V9.5a2 2 0 00-2-2h-2m-4-3H9M7 16h6M7 8h6v4H7V8z" />
-          </svg>
-        )
-      case 'performance':
-        return (
-          <svg className="w-5 h-5 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4M7.835 4.697a3.42 3.42 0 001.946-.806 3.42 3.42 0 014.438 0 3.42 3.42 0 001.946.806 3.42 3.42 0 013.138 3.138 3.42 3.42 0 00.806 1.946 3.42 3.42 0 010 4.438 3.42 3.42 0 00-.806 1.946 3.42 3.42 0 01-3.138 3.138 3.42 3.42 0 00-1.946.806 3.42 3.42 0 01-4.438 0 3.42 3.42 0 00-1.946-.806 3.42 3.42 0 01-3.138-3.138 3.42 3.42 0 00-.806-1.946 3.42 3.42 0 010-4.438 3.42 3.42 0 00.806-1.946 3.42 3.42 0 013.138-3.138z" />
-          </svg>
-        )
-    }
-  }
-
-  const getSeverityColor = (severity: Alert['severity']) => {
-    switch (severity) {
-      case 'high': return 'border-l-destructive bg-destructive/5'
-      case 'medium': return 'border-l-orange-500 bg-orange-500/5'
-      case 'low': return 'border-l-blue-500 bg-blue-500/5'
-    }
-  }
-
-  const formatTimestamp = (timestamp: Date) => {
-    const now = new Date()
-    const diff = now.getTime() - timestamp.getTime()
-    const minutes = Math.floor(diff / (1000 * 60))
-    const hours = Math.floor(diff / (1000 * 60 * 60))
-    const days = Math.floor(diff / (1000 * 60 * 60 * 24))
-
-    if (days > 0) return `${days}d ago`
-    if (hours > 0) return `${hours}h ago`
-    if (minutes > 0) return `${minutes}m ago`
-    return 'Just now'
-  }
 
   const dismissAlert = (id: string) => {
     setAlerts(prev => prev.filter(alert => alert.id !== id))
