+++ frontend/src/pages/Alterts/interfaces.ts
@@ -1,0 +1,9 @@
+export type Alert = {
+    id: string
+    type: 'risk' | 'opportunity' | 'news' | 'performance'
+    title: string
+    message: string
+    timestamp: Date
+    isRead: boolean
+    severity: 'low' | 'medium' | 'high'
+  }
