+++ frontend/src/pages/Portfolio/Portfolio.tsx
@@ -1,40 +1,7 @@
 import { useTranslation } from 'react-i18next'
 import { Navigation } from '@/components/Navigation'
 import { PieChart, Pie, Cell, ResponsiveContainer, BarChart, Bar, XAxis, YAxis, Tooltip } from 'recharts'
-
-const portfolioData = [
-  { name: 'Apple Inc.', symbol: 'AAPL', value: 45000, allocation: 32, risk: 'Medium', type: 'Stock' },
-  { name: 'Microsoft Corp.', symbol: 'MSFT', value: 38000, allocation: 27, risk: 'Low', type: 'Stock' },
-  { name: 'Tesla Inc.', symbol: 'TSLA', value: 22000, allocation: 16, risk: 'High', type: 'Stock' },
-  { name: 'S&P 500 ETF', symbol: 'SPY', value: 18000, allocation: 13, risk: 'Low', type: 'ETF' },
-  { name: 'Bitcoin', symbol: 'BTC', value: 12000, allocation: 9, risk: 'High', type: 'Crypto' },
-  { name: 'Cash', symbol: 'USD', value: 5000, allocation: 3, risk: 'None', type: 'Cash' },
-]
-
-const allocationData = [
-  { name: 'Stocks', value: 105000, color: 'hsl(var(--primary))' },
-  { name: 'ETFs', value: 18000, color: 'hsl(var(--secondary))' },
-  { name: 'Crypto', value: 12000, color: 'hsl(var(--accent))' },
-  { name: 'Cash', value: 5000, color: 'hsl(var(--muted-foreground))' },
-]
-
-const chartColors = [
-  'hsl(var(--primary))',
-  'hsl(var(--secondary))', 
-  'hsl(var(--accent))',
-  'hsl(var(--muted-foreground))',
-  'hsl(var(--primary) / 0.7)',
-  'hsl(var(--secondary) / 0.7)',
-]
-
-const performanceData = [
-  { month: 'Jan', value: 125000 },
-  { month: 'Feb', value: 132000 },
-  { month: 'Mar', value: 128000 },
-  { month: 'Apr', value: 135000 },
-  { month: 'May', value: 140000 },
-  { month: 'Jun', value: 140000 },
-]
+import { allocationData, performanceData, portfolioData } from './mocks'
 
 export default function Portfolio() {
   const { t } = useTranslation()
